Fix backup key and filter field being autofilled as logins
Firefox misinterprets the backup key field on the settings page
and the room filter field as a pair of username-password login
fields. It then autofills a username in the filter field each
time the settings page is opened.

The only working fix is to mark both fields as "autocomplete: new-password"
because Firefox ignores all other directives. I still added the proper
attributes to both fields just in case the major browsers will decide
to follow the spec one day.

The "readonly" hack is included for completeness. It may improve
compatibility with older browsers where one of the other tricks fails.

Signed-off-by: Mirian Margiani <mixosaurus+ichthyo@pm.me>